removed uncessary temp files. refactored util files to be separate files, proper names, and correct imports
