+++ backend/db/models/booking.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => {
   class Booking extends Model {
     static associate(models) {
       Booking.belongsTo(models.User, {
-        foreignKey: 'userId'
+        foreignKey: 'userId',
       })
       Booking.belongsTo(models.Spot, {
         foreignKey: 'spotId'
