+++ backend/routes/api/bookings.js
@@ -1,31 +1,14 @@
 const express = require('express')
 const router = express.Router();
 
-const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validations')
-
-const { setTokenCookie, requireAuth, restoreUser, authenticationRequired, authorizationNotRequiredBookings, authorizationRequiredBookings } = require('../../utils/auth');
 const { User, Spot, Booking } = require('../../db/models');
+const { check } = require('express-validator');
 
+const { handleValidationErrors, validateBooking } = require('../../utils/validations')
+const { restoreUser, authRequired } = require("../../utils/authentication.js");
+const { authorizationNotRequiredBookings, bookingAuthorization } = require("../../utils/authorization")
+const { validationError, notFound } = require('../../utils/helper.js')
 
-// Helper function for validation error
-const validationError = (message, code) => {
-    let error = new Error("Validation error");
-    error.status = code;
-    error.errors = {
-        endDate: message,
-        statusCode: code
-    }
-    return error;
-}
-
-// helper function for a particular element not found
-const notFound = (el, code) => {
-    let error = new Error(`${el} couldn't be found`);
-    error.status = code;
-    error.statusCode = code;
-    return error
-}
 
 // helper function for a review that already exists, may not need since this only occurs once?
 const bookingExists = (bookingError) => {
@@ -43,23 +26,10 @@ const bookingExists = (bookingError) => {
 }
 
 
-const validateBooking = [
-    check("startDate")
-        .exists({ checkFalsy: true })
-        .isString()
-        .withMessage("Start date is required"),
-    check("endDate")
-        .exists({ checkFalsy: true })
-        .isString()
-        .withMessage("End date is required"),
-    handleValidationErrors
-]
-
-
-// _________________________________________________________________________________________
+// ------------------------------------------------------------------------------------------------------
 
 // Get all of the Current User's Bookings
-router.get("/current", [restoreUser, authenticationRequired], async (req, res) => {
+router.get("/current", [restoreUser, authRequired], async (req, res) => {
     const Bookings = await Booking.findAll({
         where: {
             userId: req.user.id
@@ -76,7 +46,7 @@ router.get("/current", [restoreUser, authenticationRequired], async (req, res) =
 
 
 // Get all Bookings for a Spot based on the Spot's id
-router.get("/:spotId", [restoreUser, authenticationRequired], async (req, res, next) => {
+router.get("/:spotId", [restoreUser, authRequired], async (req, res, next) => {
     const spot = await Spot.findByPk(req.params.spotId);
     // send error if spot is not found
     if (!spot) {
@@ -109,7 +79,7 @@ router.get("/:spotId", [restoreUser, authenticationRequired], async (req, res, n
 
 
 // Create a Booking from a Spot based on the Spot's id
-router.post("/:spotId", [validateBooking, restoreUser, authenticationRequired, authorizationNotRequiredBookings], async (req, res, next) => {
+router.post("/:spotId", [validateBooking, restoreUser, authRequired, authorizationNotRequiredBookings], async (req, res, next) => {
     const { startDate, endDate } = req.body;
     // error response: body validation errors
     if (endDate <= startDate) {
@@ -149,7 +119,7 @@ router.post("/:spotId", [validateBooking, restoreUser, authenticationRequired, a
 
 
 // Edit a Booking
-router.put("/:bookingId", [restoreUser, authenticationRequired, authorizationRequiredBookings], async (req, res, next) => {
+router.put("/:bookingId", [restoreUser, authRequired, bookingAuthorization], async (req, res, next) => {
     const { startDate, endDate } = req.body;
     let compareCurrentDate = new Date();
     compareCurrentDate = compareCurrentDate.toISOString().split("T")[0]
@@ -201,7 +171,7 @@ router.put("/:bookingId", [restoreUser, authenticationRequired, authorizationReq
 })
 
 
-router.delete("/:bookingId", [restoreUser, authenticationRequired, authorizationRequiredBookings], async (req, res, next) => {
+router.delete("/:bookingId", [restoreUser, authRequired, bookingAuthorization], async (req, res, next) => {
     const deleteBooking = await Booking.findByPk(req.params.bookingId);
     // error if booking doesn't exist
     if (!deleteBooking) {
