+++ backend/routes/api/index.js
@@ -1,6 +1,6 @@
 // backend/routes/api/index.js
 const router = require("express").Router()
-const { restoreUser } = require("../../utils/auth.js");
+const { restoreUser } = require("../../utils/authentication.js");
 
 // Connect restoreUser middleware to the API router
 // If current user session is valid, set req.user to the user in the database
@@ -32,8 +32,8 @@ router.use((error, req, res, next) => {
 })
 
 
-router.post('/test', (req, res) => {
-  res.json({ requestBody: req.body });
+router.get('/test', (req, res) => {
+    res.json({message: "hello lairbnb"})
 });
 
 module.exports = router;
