+++ backend/utils/authentication.js
@@ -1,0 +1,82 @@
+// backend/utils/auth.js
+const jwt = require('jsonwebtoken');
+const { jwtConfig } = require('../config');
+const { User, Spot, Image, Review, Booking } = require('../db/models');
+const { notFound, forbidden } = require('./helper')
+
+const { secret, expiresIn } = jwtConfig;
+
+
+// Sends a JWT Cookie
+const setTokenCookie = (res, user) => {
+    // Create the token.
+    const safeUser = {
+        id: user.id,
+        email: user.email,
+        username: user.username,
+    };
+    const token = jwt.sign(
+        { data: safeUser },
+        secret,
+        { expiresIn: parseInt(expiresIn) } // 604,800 seconds = 1 week
+    );
+
+    const isProduction = process.env.NODE_ENV === "production";
+
+    // Set the token cookie
+    res.cookie('token', token, {
+        maxAge: expiresIn * 1000, // maxAge in milliseconds
+        httpOnly: true,
+        secure: isProduction,
+        sameSite: isProduction && "Lax"
+    });
+
+    return token;
+};
+
+
+// Restore user
+const restoreUser = (req, res, next) => {
+    // token parsed from cookies
+    const { token } = req.cookies;
+    req.user = null;
+
+    return jwt.verify(token, secret, null, async (err, jwtPayload) => {
+        if (err) {
+            return next();
+        }
+
+        try {
+            const { id } = jwtPayload.data;
+            req.user = await User.findByPk(id, {
+                attributes: {
+                    include: ['email', 'createdAt', 'updatedAt']
+                }
+            });
+        } catch (e) {
+            res.clearCookie('token');
+            return next();
+        }
+
+        if (!req.user) res.clearCookie('token');
+
+        return next();
+    });
+};
+
+// If there is no current user, return an error
+const authRequired = function (req, res, next) {
+    if (req.user) return next();
+
+    const error = new Error("Authentication required");
+    err.title = 'Authentication required';
+    err.errors = { message: 'Authentication required' };
+    error.status = 401
+    return next(error);
+}
+
+module.exports = {
+    setTokenCookie,
+    restoreUser,
+    authRequired,
+};
